Moved classes into separate file
